% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advRlab4package.R
\docType{package}
\name{advRlab4package-package}
\alias{advRlab4package}
\alias{advRlab4package-package}
\title{Package purpose}
\description{
Package description.
}
\author{
\strong{Maintainer}: Simon Jorstedt \email{jorstedtsimon@gmail.com}

Authors:
\itemize{
  \item 
}

}
