% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Linear regression}
\source{
\href{https://en.wikipedia.org/wiki/Linear_regression}{Linear regression on Wikipedia.}
}
\arguments{
\item{formula}{A \code{formula} object.}

\item{data}{A dataframe.}
}
\value{
An object of (RC) class \code{linreg}.
}
\description{
An RC class constructor of "linreg", for performing various linear regression tasks.
}

