% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear regression}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{A \code{formula} object.}

\item{data}{A dataframe.}
}
\value{
An object of (S3) class \code{linreg}.
}
\description{
A function for performing various linear regression tasks.
}
