% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{my_theme}
\alias{my_theme}
\title{LiU Theme}
\usage{
my_theme(plot)
}
\arguments{
\item{plot}{A \code{ggplot} object.}
}
\value{
A \code{ggplot} object.
}
\description{
A function that adds a LiU theme onto a ggplot object.
}
