% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme}
\alias{theme}
\title{LiU Theme}
\usage{
theme(data)
}
\arguments{
\item{plot}{A \code{ggplot} object.}
}
\value{
A \code{ggplot} object.
}
\description{
A function that adds a LiU theme onto a ggplot object.
}
