% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advRlab4package-package.R
\docType{package}
\name{advRlab4package-package}
\alias{advRlab4package}
\alias{advRlab4package-package}
\title{advRlab4package: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Simon Jorstedt \email{jorstedtsimon@gmail.com}

Authors:
\itemize{
  \item Muditha Cherangani \email{mcherangan@gmail.com}
}

}
\keyword{internal}
